requirements.txt
fastapi==0.110.0
passlib==1.7.4
pydantic==1.10.14
python_jose==3.3.0
replit==3.6.1
SQLAlchemy==2.0.0
uvicorn==0.20.0
email_validator==2.1.1
psycopg2-binary==2.9.5
alembic==1.13.1
PyJWT==2.6.0
bcrypt==4.0.1
python-multipart==0.0.9